Update GeneralApplicationSelectFee tests
