+++ src/main/java/uk/gov/hmcts/divorce/solicitor/event/page/GeneralApplicationSelectFee.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.divorce.solicitor.event.page;
 
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
@@ -9,11 +8,18 @@ import uk.gov.hmcts.ccd.sdk.type.DynamicList;
 import uk.gov.hmcts.divorce.common.ccd.CcdPageConfiguration;
 import uk.gov.hmcts.divorce.common.ccd.PageBuilder;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.divorcecase.model.FeeDetails;
 import uk.gov.hmcts.divorce.divorcecase.model.GeneralApplication;
+import uk.gov.hmcts.divorce.divorcecase.model.GeneralApplicationFee;
 import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.payment.service.PaymentService;
 import uk.gov.hmcts.divorce.solicitor.client.pba.PbaService;
-import uk.gov.hmcts.divorce.solicitor.service.CcdAccessService;
+
+import static uk.gov.hmcts.divorce.divorcecase.model.GeneralApplicationFee.FEE0227;
+import static uk.gov.hmcts.divorce.payment.service.PaymentService.EVENT_GENERAL;
+import static uk.gov.hmcts.divorce.payment.service.PaymentService.KEYWORD_NOTICE;
+import static uk.gov.hmcts.divorce.payment.service.PaymentService.KEYWORD_WITHOUT_NOTICE;
+import static uk.gov.hmcts.divorce.payment.service.PaymentService.SERVICE_OTHER;
 
 @Component
 @RequiredArgsConstructor
@@ -22,10 +28,6 @@ public class GeneralApplicationSelectFee implements CcdPageConfiguration {
     private final PaymentService paymentService;
 
     private final PbaService pbaService;
-
-    private final CcdAccessService ccdAccessService;
-
-    private final HttpServletRequest httpServletRequest;
 
     @Override
     public void addTo(final PageBuilder pageBuilder) {
@@ -49,8 +51,22 @@ public class GeneralApplicationSelectFee implements CcdPageConfiguration {
 
         generalApplication.getGeneralApplicationFee().setPbaNumbers(pbaNumbersDynamicList);
 
+        prepareOrderSummary(caseData);
+
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
             .data(caseData)
             .build();
+    }
+
+    private void prepareOrderSummary(CaseData data) {
+        GeneralApplicationFee feeType = data.getGeneralApplication().getGeneralApplicationFeeType();
+        FeeDetails feeDetails = data.getGeneralApplication().getGeneralApplicationFee();
+
+        String keyword = FEE0227.getLabel().equals(feeType.getLabel())
+            ? KEYWORD_NOTICE
+            : KEYWORD_WITHOUT_NOTICE;
+
+        var orderSummary = paymentService.getOrderSummaryByServiceEvent(SERVICE_OTHER, EVENT_GENERAL, keyword);
+        feeDetails.setOrderSummary(orderSummary);
     }
 }
