+++ src/test/java/uk/gov/hmcts/divorce/solicitor/event/page/GeneralApplicationSelectFeeTest.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.divorce.solicitor.event.page;
 
-import jakarta.servlet.http.HttpServletRequest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -15,7 +14,6 @@ import uk.gov.hmcts.divorce.divorcecase.model.GeneralApplication;
 import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.payment.service.PaymentService;
 import uk.gov.hmcts.divorce.solicitor.client.pba.PbaService;
-import uk.gov.hmcts.divorce.solicitor.service.CcdAccessService;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
@@ -27,10 +25,7 @@ import static uk.gov.hmcts.divorce.payment.service.PaymentService.EVENT_GENERAL;
 import static uk.gov.hmcts.divorce.payment.service.PaymentService.KEYWORD_NOTICE;
 import static uk.gov.hmcts.divorce.payment.service.PaymentService.KEYWORD_WITHOUT_NOTICE;
 import static uk.gov.hmcts.divorce.payment.service.PaymentService.SERVICE_OTHER;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.APP_1_SOL_AUTH_TOKEN;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.AUTHORIZATION;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_CASE_ID;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_SERVICE_REFERENCE;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.caseData;
 
 @ExtendWith(MockitoExtension.class)
@@ -41,12 +36,6 @@ public class GeneralApplicationSelectFeeTest {
 
     @Mock
     private PbaService pbaService;
-
-    @Mock
-    private CcdAccessService ccdAccessService;
-
-    @Mock
-    private HttpServletRequest servletRequest;
 
     @InjectMocks
     private GeneralApplicationSelectFee page;
@@ -65,7 +54,7 @@ public class GeneralApplicationSelectFeeTest {
 
         final OrderSummary orderSummary = OrderSummary.builder().build();
 
-        stubServiceRequestAndOrderSummaryCreation(orderSummary, KEYWORD_NOTICE, applicantName);
+        stubOrderSummaryCreation(orderSummary, KEYWORD_NOTICE);
 
         AboutToStartOrSubmitResponse<CaseData, State> response = page.midEvent(details, details);
 
@@ -73,10 +62,6 @@ public class GeneralApplicationSelectFeeTest {
         assertEquals(
             response.getData().getGeneralApplication().getGeneralApplicationFee().getOrderSummary(),
             orderSummary
-        );
-        assertEquals(
-            response.getData().getGeneralApplication().getGeneralApplicationFee().getServiceRequestReference(),
-            TEST_SERVICE_REFERENCE
         );
     }
 
@@ -94,7 +79,7 @@ public class GeneralApplicationSelectFeeTest {
 
         final OrderSummary orderSummary = OrderSummary.builder().build();
 
-        stubServiceRequestAndOrderSummaryCreation(orderSummary, KEYWORD_WITHOUT_NOTICE, applicantName);
+        stubOrderSummaryCreation(orderSummary, KEYWORD_WITHOUT_NOTICE);
 
         AboutToStartOrSubmitResponse<CaseData, State> response = page.midEvent(details, details);
 
@@ -102,10 +87,6 @@ public class GeneralApplicationSelectFeeTest {
         assertEquals(
             response.getData().getGeneralApplication().getGeneralApplicationFee().getOrderSummary(),
             orderSummary
-        );
-        assertEquals(
-            response.getData().getGeneralApplication().getGeneralApplicationFee().getServiceRequestReference(),
-            TEST_SERVICE_REFERENCE
         );
     }
 
@@ -133,12 +114,8 @@ public class GeneralApplicationSelectFeeTest {
         );
     }
 
-    private void stubServiceRequestAndOrderSummaryCreation(OrderSummary orderSummary, String keyword, String applicantName) {
+    private void stubOrderSummaryCreation(OrderSummary orderSummary, String keyword) {
         when(paymentService.getOrderSummaryByServiceEvent(SERVICE_OTHER, EVENT_GENERAL, keyword))
             .thenReturn(orderSummary);
-        when(servletRequest.getHeader(AUTHORIZATION)).thenReturn(APP_1_SOL_AUTH_TOKEN);
-        when(ccdAccessService.isApplicant1(APP_1_SOL_AUTH_TOKEN, TEST_CASE_ID)).thenReturn(true);
-        when(paymentService.createServiceRequestReference(null, TEST_CASE_ID, applicantName, orderSummary))
-            .thenReturn(TEST_SERVICE_REFERENCE);
     }
 }
